split box files, implement columns
