+++ foxlin/__init__.py
@@ -1,4 +1,4 @@
 from foxlin.fox import FoxLin, BASIC_BOX
-from foxlin.philosophy import Schema, Column
+from foxlin.sophy import Schema, Column
 
 __all__ = ('FoxLin', 'BASIC_BOX', 'Schema','Column')
